requirements.txt
livekit-agents
livekit-plugins-bey
livekit-plugins-openai
livekit-plugins-deepgram
livekit-plugins-silero
livekit-plugins-noise-cancellation
livekit-plugins-turn-detector

python-dotenv

fastapi
uvicorn